+++ src/routes/graphql/ts-types/post.ts
