+++ src/routes/graphql/ts-types/uuid.ts
